feat(icons): Add graphics folder, refactor graphics svg elements and add athena's icons assets
