+++ src/sections/home/collaboratorsMentors/CollaboratorsMentors.tsx
@@ -1,10 +1,10 @@
-import CollaboratorArrowOrange from '@/assets/images/icons/icon_collaboratorOrange.svg'
-import DiamondImage from '@/assets/images/icons/icon_diamond.svg'
-import MentorArrowPink from '@/assets/images/icons/icon_mentorPink.svg'
-import PurpleRectangle from '@/assets/images/icons/icon_purpleRectangle.svg'
-import PurpleTrianglesIcon from '@/assets/images/icons/icon_purpleTriangles.svg'
-import SmileIcon from '@/assets/images/icons/icon_smile.svg'
-import SwirlArrowImage from '@/assets/images/icons/icon_swirlArrow.svg'
+import CollaboratorArrowOrange from '@/assets/images/graphics/icon_collaboratorOrange.svg'
+import DiamondImage from '@/assets/images/graphics/icon_diamond.svg'
+import MentorArrowPink from '@/assets/images/graphics/icon_mentorPink.svg'
+import PurpleRectangle from '@/assets/images/graphics/icon_purpleRectangle.svg'
+import PurpleTrianglesIcon from '@/assets/images/graphics/icon_purpleTriangles.svg'
+import SmileIcon from '@/assets/images/graphics/icon_smile.svg'
+import SwirlArrowImage from '@/assets/images/graphics/icon_swirlArrow.svg'
 import CollaboratorPlaceholder from '@/assets/images/photos/photo_collaboratorPlaceholder.webp'
 import CollaboratorComputerClockImage from '@/assets/images/photos/photo_computerClockPlaceholder.webp'
 import CollaboratorComputerImage from '@/assets/images/photos/photo_computerPlaceholder.webp'
